requirements.txt
streamlit
tensorflow
pillow
numpy
streamlit-drawable-canvas